Model/requirements.txt
numpy>=1.22
Pillow>=9.0
flask>=2.0
flask-cors>=3.0.10
h5py>=3.1.0
gunicorn>=20.1.0
tensorflow>=2.10.0